Emails: Add names and logos to In-Depth Comparison page (#59717)
* Emails: Add email provider name and logo to In-Depth Comparison page

* Emails: Format email provider name and logo in In-Depth Comparison page

* Emails: Disable linting error for CSS class names

* Emails: Rename header to name in list of email provider features

* Emails: Use more descriptive class names for logos

* Emails: Fix wrong type for email provider names

* Emails: Fix TypeScript linting error
