+++ client/my-sites/email/email-providers-in-depth-comparison/data.tsx
@@ -1,11 +1,18 @@
+/* eslint-disable wpcalypso/jsx-classname-namespace */
+
+import { Gridicon } from '@automattic/components';
 import { translate } from 'i18n-calypso';
-import { EmailProviderFeatures } from 'calypso/my-sites/email/email-providers-in-depth-comparison/comparison-table';
+import googleWorkspaceIcon from 'calypso/assets/images/email-providers/google-workspace/icon.svg';
+import { getGoogleMailServiceFamily } from 'calypso/lib/gsuite';
+import { getTitanProductName } from 'calypso/lib/titan';
+import type { EmailProviderFeatures } from 'calypso/my-sites/email/email-providers-in-depth-comparison/types';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-function
 const noop = () => {};
 
 export const professionalEmailFeatures: EmailProviderFeatures = {
-	header: 'Professional Email',
+	name: getTitanProductName(),
+	logo: <Gridicon className="professional-email-logo" icon="my-sites" />,
 	tools: translate( 'Integrated email management, Inbox, calendar and contacts' ),
 	storage: translate( '30GB storage' ),
 	importing: translate( 'One-click import of existing emails and contacts' ),
@@ -14,7 +21,14 @@ export const professionalEmailFeatures: EmailProviderFeatures = {
 };
 
 export const googleWorkspaceFeatures: EmailProviderFeatures = {
-	header: 'Google Workspace',
+	name: getGoogleMailServiceFamily(),
+	logo: (
+		<img
+			alt={ translate( 'Google Workspace icon', { textOnly: true } ) }
+			className="google-workspace-logo"
+			src={ googleWorkspaceIcon }
+		/>
+	),
 	tools: translate( 'Gmail, Calendar, Meet, Chat, Drive, Docs, Sheets, Slides and more' ),
 	storage: translate( '30GB storage' ),
 	importing: translate( 'Easy to import your existing emails and contacts' ),
